server/requirements.txt
adium-theme-ubuntu==0.3.4
attrs==15.2.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
boto==2.38.0
chainer==1.13.0
chardet==2.3.0
click==6.6
cryptography==1.2.3
decorator==4.0.10
duplicity==0.7.6
enum34==1.1.2
filelock==2.0.6
Flask==0.11.1
Flask-Cors==2.1.2
html5lib==0.999
idna==2.0
ImageHash==2.2
imageio==1.5
ipaddress==1.0.16
ipdb==0.10.1
ipython==5.0.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
lockfile==0.12.2
lxml==3.5.0
MarkupSafe==0.23
moviepy==0.2.2.11
ndg-httpsclient==0.4.0
nltk==3.2.1
nose==1.3.7
numpy==1.11.1
PAM==0.4.2
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.3
Pillow==3.3.0
prompt-toolkit==1.0.3
protobuf==3.0.0
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycurl==7.43.0
Pygments==2.1.3
pygobject==3.20.0
pygpgme==0.3
pyOpenSSL==0.15.1
pyserial==3.0.1
Pyste==0.9.10
python-apt==1.1.0b1
python-cloudfiles==1.7.10
PyWavelets==0.4.0
pyxdg==0.25
requests==2.9.1
scipy==0.18.0
service-identity==16.0.0
simplegeneric==0.8.1
six==1.10.0
tensorflow==0.9.0
tqdm==4.8.1
traitlets==4.2.2
Twisted==16.0.0
unity-lens-photos==1.0
urllib3==1.13.1
virtualenv==15.0.2
wcwidth==0.1.7
Werkzeug==0.11.10
zope.interface==4.1.3
